refactor: isolate template clause link operations
