+++ apps/api/src/handlers/docx/resolve-clause-slots.ts
@@ -10,7 +10,7 @@ import {
   clauseBodyToPlainText,
   clauseBodyToRichPatch,
 } from "@/api/handlers/clauses/clause-to-patch";
-import { isVariantDeleted } from "@/api/handlers/clauses/template-links";
+import { isVariantDeleted } from "@/api/lib/template-clause-links";
 import type { ClauseBody } from "@/api/handlers/clauses/types";
 import type { SafeId } from "@/api/lib/branded-types";
