+++ apps/api/src/handlers/templates/check.ts
@@ -2,7 +2,7 @@ import { Result } from "better-result";
 import { t } from "elysia";
 
 import type { SafeDb, ScopedDb } from "@/api/db/safe-db";
-import { listTemplateClausesHandler } from "@/api/handlers/clauses/template-links";
+import { listTemplateClausesHandler } from "@/api/lib/template-clause-links";
 import { discoverClauseSlots } from "@/api/handlers/docx/discover-clause-slots";
 import { discoverTemplate } from "@/api/handlers/docx/discover-template";
 import { extractText } from "@/api/handlers/docx/extract-text";
